Expose 'null' match and fwd/reverse scores in sample
When an attribute exists in the item that we want to match against
but not in the target, for example someone has not provided the
full details of their car, house etc, then we have a choice as to
how to deal with it.

By default, we score zero for 'null' and then apply the attribute
weighting (thus a weighting of 0.7 would give a score of 30%).

For 'strict' weighting (i.e. 1.0), the score would be zero and
would result in the overall score being zero (as it's a geometric
mean).

Alternatively, we can tell the matching engine to assume a certain
level of relevance - effectively taking a guess as to an average
likely match.

This can now be seen by looking at the matches against the record
named 'Null'